wpan-ping: Ping utility on IEEE 802.15.4 level
wpan-ping aims to offer ping/ping6 like functionality on IEEE 802.15.4 level.

No control message protocol is defined so we will simply use DGRAM's over a
plain socket with a server component to emulate the behaviour. Right now it
is possible to specify packet count as well as size and the interface to be
used. So far it uses short addresses only. Support for extended addresses is
on the todo list.

Signed-off-by: Stefan Schmidt <stefan@osg.samsung.com>
Signed-off-by: Alexander Aring <alex.aring@gmail.com>