wpan-ping: Take 6LoWPAN dispatch byte into account
In rfc4944 section 5.1 a dispatch byte is defined as the first transmitted
byte in the payload. All protocols that want to work in 6LoWPAN networks
without interfering with them should set it to something like 00xxxxxx.

We want to operate in such networks without disturbing them so putting a safe
0x00 here. We might use some of the other bits for flags at some point but do
not need them right now.

And yes, the maximum payload is still increasing by one instead of
decreasing. Earlier versions of this tool had a longer internal header which
resulted in a smaller payload length. That has changed now so we can increase
the payload length by 2 and subtract the dispatch byte to still gain one byte.

Signed-off-by: Stefan Schmidt <stefan@osg.samsung.com>
Signed-off-by: Alexander Aring <alex.aring@gmail.com>