wpan-ping: unify src and dst address handling
Instead of having parts of the sockaddr struct parsed and kept in different
places we put it all in the correct struct from the start now. All local
information for the src address are requested over netlink while the dst
address is parsed from the commandline. As the PAN ID has to be the same we
take this also from netlink for dst.

Signed-off-by: Stefan Schmidt <stefan@osg.samsung.com>
Signed-off-by: Alexander Aring <alex.aring@gmail.com>