wpan-ping: do not try to send data back to origin if we got an error from recvfrom
In the case of a length of -1 from recvfrom we print out the error code but we
never left the current loop iteration. That might be problematic when we try to
send the data back with sendto. Instead we now continue with the next iteration
from now on if we get an error.

CID: 32115
Signed-off-by: Stefan Schmidt <stefan@osg.samsung.com>