ci build: add Coverity Scan service support to pur Travis config
This plugin allows us to submit builds to the Coverity Scan service for static
analysis of our code base. We are not submitting every push to master to
Coverity as this would be overkill. Instead a push to a branch named
coverity_scan will submit it.

This also allows to test your code before pushing it into the master branch.
If you want to get added to the wpan-tools project at Coverity Scan service
go here, create an account and request to get added:

https://scan.coverity.com/projects/linux-wpan-wpan-tools/

Signed-off-by: Stefan Schmidt <stefan@osg.samsung.com>