wpan-ping: set more default values for our configuration
Make sure we actually have sane default values for the internal
configuration. This also fixes a side effect from the calloc fix.
The packet counter was zero from now on but we actually want the
highest number when no other value is given on the commandline.

Also fix a typo while being there.

Signed-off-by: Stefan Schmidt <stefan@osg.samsung.com>