src/nl_extras.h: fix compatibility with libnl 3.3.0
nl_extras.h defines a set of nla_set_s*() functions if not provided by
libnl. They are provided by libnl since version 3.2.26. The test
(LIBNL_VER_MIC <= 26) was working fine while libnl was in the 3.2.x
series, but now that they have incremented the minor version, the
micro version was reset to 0, with the latest libnl version being
3.3.0.

Due to this, the condition (LIBNL_VER_MIC <= 26) is true, and we get
redefinition errors because nl_extras.h redefines functions already
provided by libnl.

This commit improves the condition so that nl_extras.h provides the
missing functions only if the minor version is < 2, or if minor is 2
and micro is < 26.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>